1. 表单定义表格列调整，去掉后处理列，编辑器列放置在第一列，类型名称--改为:文档类型名称 2. 表单定义表格组件的查询，排序功能不正确。 3. 表单定义的删除服务，需要连带删除 文档导出规则 4. 创建文档到处规则对话框调整，将后处理脚本 作为单独一页。将字段放置于第一页的文档编辑器栏位下方。第一页页签改名为导出字段 5. 导出字段表格尺寸适度缩小，避免编辑单元格时跳动 6. 对象选择单元格，必须有内容，能显示选择的对象 7. 下方的，创建，删除，清空等按钮右移动8px，不要左贴边。 8. 表格增加勾选框，删除的是删除勾选项 9. 清空提示.

+++ com.bizvisionsoft.pms/src/com/bizvisionsoft/pms/formdef/action/FormDefACT.java
@@ -1,9 +1,13 @@
 package com.bizvisionsoft.pms.formdef.action;
 
+import java.util.Optional;
+
 import org.eclipse.jface.dialogs.IDialogConstants;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.nebula.jface.gridviewer.GridTreeViewer;
 import org.eclipse.swt.widgets.Event;
 
+import com.bizivisionsoft.widgets.util.Layer;
 import com.bizvisionsoft.annotations.AUtil;
 import com.bizvisionsoft.annotations.md.service.Behavior;
 import com.bizvisionsoft.annotations.ui.common.Execute;
@@ -60,9 +64,9 @@ public class FormDefACT {
 	 * @param element
 	 */
 	private void doCreate(Object element) {
-		ExportDocRule exportDocRule = ((FormDef) element).newSubItem();
-		if (IDialogConstants.OK_ID == EditExportDocRuleDialog.open(br, context, exportDocRule)) {
-			service.insertExportDocRule(exportDocRule, br.getDomain());
+		EditExportDocRuleDialog dialog = EditExportDocRuleDialog.create(br, context, ((FormDef) element).newSubItem());
+		if (IDialogConstants.OK_ID == dialog.open()) {
+			service.insertExportDocRule(dialog.getExportDocRule(), br.getDomain());
 			((IQueryEnable) context.getContent()).doRefresh();
 		}
 	}
@@ -73,12 +77,17 @@ public class FormDefACT {
 	 * @param element
 	 */
 	private void doDelete(Object element) {
-		if (element instanceof FormDef)
-			service.deleteFormDef(((FormDef) element).get_id(), br.getDomain());
-		else
-			service.deleteExportDocRule(((ExportDocRule) element).get_id(), br.getDomain());
+		String label = AUtil.readTypeAndLabel(element);
+		String message = Optional.ofNullable(label).map(m -> "请确认将要删除 " + m).orElse("请确认将要删除选择的记录。");
+		if (MessageDialog.openConfirm(br.getCurrentShell(), "删除", message)) {
+			if (element instanceof FormDef)
+				service.deleteFormDef(((FormDef) element).get_id(), br.getDomain());
+			else
+				service.deleteExportDocRule(((ExportDocRule) element).get_id(), br.getDomain());
 
-		((IQueryEnable) context.getContent()).doRefresh();
+			((IQueryEnable) context.getContent()).doRefresh();
+			Layer.message("已删除");
+		}
 	}
 
 	/**
@@ -94,10 +103,12 @@ public class FormDefACT {
 				viewer.update(AUtil.simpleCopy(d, element), null);
 			});
 		else if (element instanceof ExportDocRule) {
-			if (IDialogConstants.OK_ID == EditExportDocRuleDialog.open(br, context, (ExportDocRule) element)) {
+			EditExportDocRuleDialog dialog = EditExportDocRuleDialog.create(br, context, (ExportDocRule) element);
+			if (IDialogConstants.OK_ID == dialog.open()) {
+				ExportDocRule exportDocRule = dialog.getExportDocRule();
 				service.updateExportDocRule(new FilterAndUpdate().filter(new BasicDBObject("_id", ((ExportDocRule) element).get_id()))
-						.set(BsonTools.getBasicDBObject((ExportDocRule) element, "_id")).bson(), br.getDomain());
-				viewer.update(element, null);
+						.set(BsonTools.getBasicDBObject(exportDocRule, "_id")).bson(), br.getDomain());
+				viewer.update(AUtil.simpleCopy(exportDocRule, element), null);
 			}
 		}
